refactored presentation
