+++ src/controller/UserController/index.ts
@@ -1,60 +1,23 @@
 import { getRepository } from "typeorm";
-import { NextFunction, Request, Response } from "express";
 import { User } from "../../entity/User";
-import { validateOrReject } from "class-validator";
-import { SaveUserInput } from "./input/save";
-import { v4 } from "uuid";
-import { GetUserByIdInput } from "./input/getUserById";
-import { DeleteUserInput } from "./input/deleteUser";
 
 export class UserController {
   private userRepository = getRepository(User);
 
-  async all(request: Request, response: Response, next: NextFunction) {
-    const users = (await this.userRepository.find()) || [];
-    return {
-      users,
-    };
+  public async all() {
+    return (await this.userRepository.find()) || [];
   }
 
-  async one(request: Request, response: Response, next: NextFunction) {
-    const id = request.params.id;
-
-    const input = new GetUserByIdInput();
-    input.id = id;
-
-    validateOrReject(input);
-
-    return this.userRepository.findOne(input.id);
+  async one(id: string) {
+    return this.userRepository.findOne(id);
   }
 
-  async save(request: Request, response: Response, next: NextFunction) {
-    const input = new SaveUserInput();
-
-    const id = v4();
-
-    input.name = request.body.name;
-    input.nickname = request.body.nickname;
-    input.email = request.body.email;
-
-    validateOrReject(input);
-
-    return this.userRepository.save({
-      id,
-      name: input.name,
-      nickname: input.nickname,
-      email: input.email,
-    });
+  async save(input: any) {
+    return this.userRepository.save(input);
   }
 
-  async remove(request: Request, response: Response, next: NextFunction) {
-    const input = new DeleteUserInput();
-
-    input.id = request.params.id;
-
-    validateOrReject(input);
-
-    let userToRemove = await this.userRepository.findOne(request.params.id);
+  async remove(id: string) {
+    let userToRemove = await this.userRepository.findOne(id);
 
     if (userToRemove) {
       await this.userRepository.remove(userToRemove);
