--- src/controller/UserController/input/deleteUser.ts
@@ -1,7 +1,0 @@
-import { IsString, IsNotEmpty } from "class-validator";
-
-export class DeleteUserInput {
-  @IsString()
-  @IsNotEmpty()
-  id?: string;
-}
