--- src/controller/UserController/input/getUserById.ts
@@ -1,7 +1,0 @@
-import { IsString, IsNotEmpty } from "class-validator";
-
-export class GetUserByIdInput {
-  @IsString()
-  @IsNotEmpty()
-  id?: string;
-}
