--- src/controller/UserController/input/save.ts
@@ -1,15 +1,0 @@
-import { IsString, IsNotEmpty } from "class-validator";
-
-export class SaveUserInput {
-  @IsString()
-  @IsNotEmpty()
-  name?: string;
-
-  @IsString()
-  @IsNotEmpty()
-  nickname?: string;
-
-  @IsString()
-  @IsNotEmpty()
-  email?: string;
-}
